Basic API changes

+++ Code/Framework/AzToolsFramework/AzToolsFramework/AssetBrowser/AssetBrowserTreeToTableProxyModel.cpp
@@ -63,7 +63,7 @@ namespace AzToolsFramework
 
         int AssetBrowserTreeToTableProxyModel::rowCount(const QModelIndex& parent) const
         {
-            return parent.isValid() ? 0 : flattenedData.size();
+            return parent.isValid() ? 0 : static_cast<int>(flattenedData.size());
         }
 
         int AssetBrowserTreeToTableProxyModel::columnCount(const QModelIndex& parent) const
